Переместит mapDispatchToProps из ArticleList в Article
